feat(inbox): desktop notifications
