+++ src/components/EnableNotifications.jsx
@@ -1,15 +1,20 @@
 import { Button } from '@ynput/ayon-react-components'
-import useNotification from '../hooks/useNotification'
+import sendNotification from '../helpers/browserNotification'
+import { useNavigate } from 'react-router'
 
 const EnableNotifications = () => {
-  const sendFirstNotification = useNotification()
+  const navigate = useNavigate()
 
   return (
     Notification.permission !== 'granted' && (
       <Button
         icon={'notifications'}
         data-tooltip="Get notifications on your device"
-        onClick={() => sendFirstNotification('Notifications already enabled 💪')}
+        onClick={() =>
+          sendNotification({ title: 'Notifications already enabled 💪' }, () =>
+            navigate('/account/profile'),
+          )
+        }
         variant="filled"
       >
         Enable notifications
